Refactor task scheduling related components
This is a preparation for introducing task level retries.

This PR removes "all-or-nothing" execution assumption from SqlStageExecution.
Scheduling related assumptions of streaming execution are now moved to
the StreamingStageExecution. SqlStageExecution is now merely a container
for scheduled tasks for a particular stage. It's main responsibility now
is to keep track of running tasks and collect their execution
statistics.